Update BuildUSBDrives.ps1
Re-wrote how the script to optimize how the script partitions and copies the files to drives using start-job. it will now start a job to partition each drive, wait for them to complete, create a new job for each drive to copy boot files, Images, and drivers. This reduced drive creation time for 3 drives from 21 minutes to ~8 minutes.