-Update Check if unattend_xx.xml or Prefixes.txt file in unattend folder exist. If not writelog not found and continues.
-Add:  If Prefixes.txt is found, it will copy to the Deploy\unattend
folder. No need to remeber to go back and manually copy the file over for computer
naming.

-could consider renaming the initial Prefixes.txt in the repository to Prefixes_Sample.txt
to be consistent with the Applist_Sample.json. Assumption is that if you have made changes to the file,
then you want to use it to cut down on the number of reboots and reduce app packages, as per Richards Youtube Presentation.